Move logical tables related classes from pinot-query-planner to pinot-core
