+++ pinot-controller/src/test/java/org/apache/pinot/controller/util/BrokerServiceHelperTest.java
@@ -25,7 +25,7 @@ import java.util.Map;
 import org.apache.helix.model.InstanceConfig;
 import org.apache.pinot.controller.ControllerConf;
 import org.apache.pinot.controller.helix.core.PinotHelixResourceManager;
-import org.apache.pinot.core.routing.TimeBoundaryInfo;
+import org.apache.pinot.core.routing.timeboundary.TimeBoundaryInfo;
 import org.apache.pinot.spi.config.table.TableConfig;
 import org.apache.pinot.spi.config.table.TableType;
 import org.apache.pinot.spi.utils.builder.TableConfigBuilder;
